Module Abt.Make

Parameters

O : Operator

Signature

type t

The type of ABT's constructed from the operators defind in O

val v : string -> t

v x is a leaf in the ABT consisting of a variable named x

val op : t O.t -> t

op o is a branch in the ABT consisting of the operator o

val (#.) : string -> t -> t

x #. t is a new abt obtained by binding all free variables named x in t

Note that this does not bind the variables to a value, (for which, see subst). This only binds the free variables within the scope of an abstraction that ranges over the given (sub) abt t.

val subst : string -> value:t -> t -> t

subst name ~value t is a new abt obtained by substituting value for the outermost scope of variables bound to name in t

val to_string : t -> string

to_string t is the representation of t as a string

val equal : t -> t -> bool

equal t t' is true when t and t' are alpha equivalent and false otherwise

val case : var:(Var.t -> 'a) -> bnd:(Var.Binding.t -> t -> 'a) -> opr:(t O.t -> 'a) -> t -> 'a

val transform : var:(Var.t -> Var.t) -> bnd:((Var.Binding.t * t) -> Var.Binding.t * t) -> opr:(t O.t -> t O.t) -> t -> t