Module Unification.Subst

type term = t

An alias for the type of the ABT for reference in the context of the substitution

type t

Substitutions mapping free variables to terms

val find : Var.t -> t -> term option

find v s is Some term if v is bound to term in the substitution s, otherwise it is None

val bindings : t -> (Var.t * term) list

bindings s is a list of all the bindings in s

val to_string : t -> string