Module Var.Binding

A binding is an immutable reference, to which a bound can refer.

include Stdlib.Map.OrderedType

type t

val compare : t -> t -> int

val equal : t -> t -> bool

val v : string -> t

binding s is a new binding for variables named s

val name : t -> string